Revisit error scenarios (#73)
Main change - if encoded or archive file already exists, then aborts so
files aren't overwritten. Returns exit code 1 if anything failed.

+++ auto-handbrake-cfr/src/main/java/com/willmolloy/handbrake/cfr/VideoArchiver.java
@@ -13,7 +13,7 @@ import org.apache.logging.log4j.Logger;
  *
  * @author <a href=https://willmolloy.com>Will Molloy</a>
  */
-public class VideoArchiver {
+class VideoArchiver {
 
   private static final Logger log = LogManager.getLogger();
 
@@ -21,37 +21,46 @@ public class VideoArchiver {
    * Archives the given video.
    *
    * @param video video to archive
+   * @return {@code true} if archiving was successful
    */
-  public CompletableFuture<Void> archiveAsync(UnencodedVideo video) {
-    return CompletableFuture.runAsync(
+  public CompletableFuture<Boolean> archiveAsync(UnencodedVideo video) {
+    return CompletableFuture.supplyAsync(
         () -> {
           Stopwatch stopwatch = Stopwatch.createStarted();
           try {
-            log.info("Archiving: {} -> {}", video.originalPath(), video.archivedPath());
+            log.debug("Archiving: {} -> {}", video.originalPath(), video.archivedPath());
 
             if (Files.exists(video.archivedPath())) {
               if (Files.mismatch(video.originalPath(), video.archivedPath()) == -1) {
-                log.warn(
-                    "Archive file ({}) already exists, deleting original", video.archivedPath());
-                Files.delete(video.originalPath());
+                log.warn("Archive file ({}) already exists", video.archivedPath());
+                if (!video.originalPath().equals(video.archivedPath())) {
+                  log.info("Deleting: {}", video.originalPath());
+                  Files.delete(video.originalPath());
+                }
+                return true;
               } else {
-                log.warn(
-                    "Archive file ({}) already exists but contents differ, keeping original",
+                log.error(
+                    "Archive file ({}) already exists but contents differ. Aborting",
                     video.archivedPath());
+                return false;
               }
-              return;
             }
 
             Files.createDirectories(checkNotNull(video.archivedPath().getParent()));
 
+            log.info("Moving: {} -> {}", video.originalPath(), video.archivedPath());
             // archive to a temp file first in case something goes wrong
             // (e.g. app crash while it's uploading to NAS)
             Files.move(video.originalPath(), video.tempArchivedPath());
             Files.move(video.tempArchivedPath(), video.archivedPath());
 
-            log.info("Archived: {} - elapsed: {}", video.archivedPath(), stopwatch.elapsed());
+            log.info("Archived: {}", video.archivedPath());
+            return true;
           } catch (Exception e) {
             log.error("Error archiving: %s".formatted(video), e);
+            return false;
+          } finally {
+            log.info("Elapsed: {}", stopwatch.elapsed());
           }
         });
   }
